add rpc interface, use general phase start and end messages, add gob_rpc, possibly did other things

+++ waffle/loader.go
@@ -6,7 +6,7 @@ import (
 
 type Loader interface {
 	Init(*Worker)
-	Load() (int, os.Error)
+	Load() (uint64, os.Error)
 }
 
 type LoaderBase struct {
